Following discussion in #2849, decided to make the new wraps implementation available on Python 3.4 as well.  As such I moved it out of astropy.utils.compat since it's really providing a new function rather than some backward/forward compatibility.
